\name{Bug_Metrics_Software}
\alias{Bug_Metrics_Software}
\docType{data}
\title{
Bug Metrics Data
}
\description{
A data set which reports the 5 different type of bugs for 5 software. The count frequencies are available for pre- and post- release of the data. 
}
\usage{data(Bug_Metrics_Software)}
\format{
  A three dimensional array on the bug counts of 5 software at 5 severity levels. 
}
\source{
http://www.eclipse.org/jdt/core/index.php
}
\examples{
data(Bug_Metrics_Software)
}
\keyword{Bar plot}

