\name{employ}
\alias{employ}
\docType{data}
\title{
A data set used for understanding the very basic steps in R}
\description{
The data set is used to simply understand the working of read.table, View, class and sapply R functions}
\usage{data(employ)}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{\code{Trade}}{a numeric vector}
    \item{\code{Food}}{a numeric vector}
    \item{\code{Metals}}{a numeric vector}
  }
}
\examples{
data(employ)
}
\keyword{Basic Tools}

