\name{Gasoline}
\alias{Gasoline}
\docType{data}
\title{
Car Mileage Dataset
}
\description{
This data set has been used primarily for understanding a multivariate data set. Multiple regression model is also introduced and discussed completely through this example.}
\usage{data(Gasoline)}
\format{
  A data frame with 25 observations on the following 12 variables.
  \describe{
    \item{\code{y}}{Miles per gallon}
    \item{\code{x1}}{Displacement (cubic inches)}
    \item{\code{x2}}{Horsepower (foot-pounds)}
    \item{\code{x3}}{Torque (foot-pounds)}
    \item{\code{x4}}{Compression ratio}
    \item{\code{x5}}{Rear axle ratio}
    \item{\code{x6}}{Carburetor (barrels)}
    \item{\code{x7}}{Number of transmission speeds}
    \item{\code{x8}}{Overall length (inches)}
    \item{\code{x9}}{Width (inches)}
    \item{\code{x10}}{Weight (pounds)}
    \item{\code{x11}}{Type of transmission (A-automatic, M-manual)}
  }
}
\references{
Montgomery, D. C., Peck, E.A., and Vining, G.G. (2012). Introduction to linear regression analysis. Wiley.}
\examples{
data(Gasoline)
}
\keyword{Linear multiple regression model}

