\name{OF}
\alias{OF}
\docType{data}
\title{
Understanding the Overfitting Problem
}
\description{
This is a data set cooked up by the author to highlight the problem of overfitting. The variables have no physical meaning. 
}
\usage{data(OF)}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{X}}{Just another covariate}
    \item{\code{Y}}{Just another output}
  }
}
\examples{
data(OF)
plot(OF)
}
\keyword{Overfitting}

