\name{PW_Illus}
\alias{PW_Illus}
\docType{data}
\title{
A data set for illustrating "Piecewise Linear Regression Model" }
\description{
As with the "OF" data set, this data set has been created by the author to build up the ideas leading up to piecewise linear regression model. 
}
\usage{data(PW_Illus)}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{X}}{an input vector}
    \item{\code{Y}}{an output vector}
  }
}
\examples{
data(PW_Illus)
plot(PW_Illus)
}
\keyword{Piece-wise Linear Regression}

