\name{RSADBE-package}
\alias{RSADBE-package}
\alias{RSADBE}
\docType{package}
\title{Data Sets for the "R Statistical Application Development by Example" Book
}
\description{
The RSADBE package contains all the data sets used in the book "R Statistical Application Development by Example". Data sets have been collected from various sources and an attempt has been made to ensure that all the right credits are given. If some omissions are there, kindly accept the current work as a compliment for your work. 
}
\details{
\tabular{ll}{
Package: \tab RSADBE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-13\cr
License: \tab GPL-2\cr
}
This package is aimed to complement the book. Any data set required in the book may simply loaded using data(GC) as an example. 
}
\author{
Prabhanjan

Maintainer: Prabhanjan Tattar <prabhanjannt@gmail.com>

}
\references{
Tattar, P.N. (2013). R Statistical Application Development by Example. Packt Publication.
}

\examples{
data(GC)
}
\keyword{ RSADBE }
