\name{SCV}
\alias{SCV}
\docType{data}
\title{
An illustrative data set where the "Response" depends on four variables A-D and a fifth categorical variable}
\description{
This data set is primarily used to illustrate some basic R functions. 
}
\usage{data(SCV)}
\format{
  A data frame with 16 observations on the following 6 variables.
  \describe{
    \item{\code{Response}}{an output vector}
    \item{\code{A}}{variable A}
    \item{\code{B}}{variable B}
    \item{\code{C}}{Variable C}
    \item{\code{D}}{variable D}
    \item{\code{E}}{a factor with two levels \code{Modified} \code{Usual}}
  }
}
\examples{
data(SCV)
}
\keyword{datasets}

