\name{SCV_Modified}
\alias{SCV_Modified}
\docType{data}
\title{
SCV data set by category "Modified"
}
\description{
This data set is a part of the SCV dataset. 
}
\usage{data(SCV_Modified)}
\format{
  A data frame with 8 observations on the following 6 variables.
  \describe{
    \item{\code{Response}}{an output vector}
    \item{\code{A}}{variable A}
    \item{\code{B}}{variable B}
    \item{\code{C}}{Variable C}
    \item{\code{D}}{variable D}
    \item{\code{E}}{a factor with two levels \code{Modified}}
  }
}

\examples{
data(SCV_Modified)
}
\keyword{datasets}

