\name{SCV_Usual}
\alias{SCV_Usual}
\docType{data}
\title{
SCV data set with caterogy "Usual"}
\description{
This data set is part of the SCV data set. 
}
\usage{data(SCV_Usual)}
\format{
  A data frame with 8 observations on the following 6 variables.
  \describe{
  \item{\code{Response}}{an output vector}
    \item{\code{A}}{variable A}
    \item{\code{B}}{variable B}
    \item{\code{C}}{Variable C}
    \item{\code{D}}{variable D}
    \item{\code{E}}{a factor with two levels \code{Usual}}
    }
}
\examples{
data(SCV_Usual)
}
\keyword{datasets}

