\name{simpledata}
\alias{simpledata}
\docType{data}
\title{
A simulated data set for illustrating the ROC concept
}
\description{
ROC is an important tool for comparing different models for the same classification problem. This data set comes with barebones infrastructure and is simply complementary in nature towards setting up a clear understanding the ROC construction. 
}
\usage{data(simpledata)}
\format{
  A data frame with 200 observations on the following 2 variables.
  \describe{
    \item{\code{Predictions}}{Predicted probabilities}
    \item{\code{Label}}{True class of the observations}
  }
}
\examples{
data(simpledata)
}
\keyword{datasets}

