\name{SQ}
\alias{SQ}
\docType{data}
\title{
Sample Questionnaire Data
}
\description{
The sample questionnaire data is simply used to familiarize the reader with data and statistical terminologies. 
}
\usage{data(SQ)}
\format{
  A data frame with 20 observations on the following 12 variables.
  \describe{
    \item{\code{Customer_ID}}{Customer ID}
    \item{\code{Questionnaire_ID}}{Questionnaire ID}
    \item{\code{Name}}{Customers Name}
    \item{\code{Gender}}{Customers gender \code{Female} \code{Male}}
    \item{\code{Age}}{Age of the customer}
    \item{\code{Car_Model}}{Car Model's name}
    \item{\code{Car_Manufacture_Year}}{Month and year of car's manufacturing}
    \item{\code{Minor_Problems}}{Minor problems were fixed by the workshop center indicator \code{No} \code{Yes}}
    \item{\code{Major_Problems}}{Major problems were fixed by the workshop center indicator \code{No  } \code{Yes} \code{Yes }}
    \item{\code{Mileage}}{The overall mileage of the car (kms/litre) }
    \item{\code{Odometer}}{The overall kilometers travelled by the car}
    \item{\code{Satisfaction_Rating}}{How satisfied was the customer \code{Very Poor} < \code{Poor} < \code{Average} < \code{Good} < \code{Very Good}}
  }
}
\examples{
data(SQ)
}
\keyword{Sample Data}
